# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: MedschoolProffs learning platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Sign in
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSession"
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a student
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Student created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSession"
...
components:
...
  schemas:
...
    User:
      type: object
      required: [id, name, email, role, status]
      properties:
        id: { type: integer }
        name: { type: string }
        email: { type: string }
        role: { type: string }
        status: { type: string }
        institution: { type: ["string", "null"] }
        program: { type: ["string", "null"] }
    AuthSession:
      type: object
      required: [token, user]
      properties:
        token: { type: string }
        user: { $ref: "#/components/schemas/User" }
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 8 }
    RegisterInput:
      type: object
      required: [name, email, password, institution, program, academicYear, batch, rollNumber]
      properties:
        name: { type: string, minLength: 2 }
        email: { type: string, format: email }
        password: { type: string, minLength: 8 }
        institution: { type: string }
        program: { type: string }
        academicYear: { type: string }
        batch: { type: string }
        rollNumber: { type: string }
        phone: { type: ["string", "null"] }
...